Add licenses, add pdf meta, remove latex silence
The Midgard codebase is licensed under Apache 2.0, which is a
standard choice for code.

The Midgard technical specification is licensed under Creative
Commons BY-SA 4.0, which is standard for docs.

Defined some standard PDF metadata fields (e.g. title, author).

Removed the silence package from the latex code, as it was not
playing well with some other packages.

Also, reworked the inline code hyphenation.