Remove inkscape dependency and replace SVGs
Inkscape is an awkward dependency in Latex, requiring the arguably
unsafe `-shell-escape` command-line option and special configuration
for latex editor integration.

It's better to export PDFs via Inkscape directly whenever the SVGs
are changed. If this happens frequently, we can add a helper script
to refresh the PDF exports (when needed) as a git pre-commit hook.
Nonetheless, the source SVGs should still be version-controlled.