feat: first draft of withdrawal contract
By the time a withdrawal event can be concluded, it's been merged into
the confirmed state, which in turn has the UTxO in question removed from
its ledger. And settlement nodes only carry user event trees.

So I've added `l2_value` to `WithdrawalInfo` as there is no way to know
how much value had been contained within the withdrawn UTxO at L2.

This commit also contains a barebones library module for `reserve`, with
the assumption that it'll rely on the withdraw-0 trick to validate
spends, as it'll probably be required to allow multiple UTxOs to be
spent in single transactions.