fix: remove `Output` from witness redeemer
The underlying `Value` of `Output` is an opaque type, and it failed with
"an opaque type is trying to escape" error. By changing the type to
`Data`, and unsafely extracting its first two elements (assuming a
`Constr`, without checking its index), we can get around this and coerce
the value as our `ValuePairs` datatype.

The reason this unsafe behavior is acceptable is that we are already
validating the presence of the subject UTxO in the ledger tree using
`plutarch_phas`. So there is no need to validate the correctness of its
structure for extracting relevant data.